+++ save-sqlaudit-files/Function/SqlAuditToCloudLog.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+ <PropertyGroup>
+   <OutputType>Exe</OutputType>
+   <TargetFramework>net6.0</TargetFramework>
+ </PropertyGroup>
+ 
+ <ItemGroup>
+    <PackageReference Include="Google.Cloud.Functions.Hosting" Version="1.0.0" />
+    <PackageReference Include="Google.Events.Protobuf" Version="1.0.0" />
+    <PackageReference Include="Google.Cloud.Storage.V1" Version="4.1.0" />
+    <PackageReference Include="Google.Cloud.PubSub.V1" Version="3.2.0" />
+    <PackageReference Include="Microsoft.SqlServer.XEvent.XELite" Version="2021.12.12.2" />
+    <PackageReference Include="Google.Cloud.Logging.V2" Version="4.0.0" />
+    <PackageReference Include="Google.Cloud.Audit" Version="2.1.0" />
+ </ItemGroup>
+</Project>
