+++ save-sqlaudit-files/Run/env.yml
@@ -2,4 +2,7 @@ PROJECT_ID: ""
 LOG_ID: ""
 SETTINGS_FILENAME: "mySettings.json"
 DO_PUBSUB_PUBLISH: "False"
+DO_PARQUET_GCS_SAVE: "True"
+PARQUET_GCS_BUCKETNAME: ""
+DO_LOG_SAVE: "False"
 BATCH_SIZE: "1000"
